Create table body component
